clientes premium query de faltante de stock correcciones en el form de facturacion ( checkear sotck, limpiar datos ) item de facturas ahora vanca mas de 255 to string de los modelos de tablero

+++ visual/VentaElectrodomesticos/Modelo/ItemFactura.cs
@@ -26,7 +26,7 @@ namespace VentaElectrodomesticos.Modelo {
             set;
         }
 
-        public Byte cantidad {
+        public Int32 cantidad {
             get;
             set;
         }
